src/screen.c: Suppress set-but-unused variable warnings
Mark the variables with UNUSED() instead of removing them because
the calls may have side effects.

Although there is a commented out call that prints these
variables, i.e. scr_tsconf_putsprite_comun(). Export this
function in tsconf.h so uncommenting the call during debugging
doesn't cause an implicit declaration warning.

Remove the #elif defined(__cplusplus) part of the UNUSED() macro
in compiler.h, the whole project is written in C and removing
the variable name itself is not useful anyway, the UNUSED macro
can be used for both variables and function parameters.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>