src/menu.c: Fix a signed char comparison warning.
The value range for signed char is -128 ... 127 so the
check for >127 is always false and is covered by the <32 check.
Remove the second check. Fixes this warning:

menu.c: In function ‘menu_file_mmc_browser_show_file’:
menu.c:18053:31: error: comparison is always false due to limited range of data type [-Werror=type-limits]
    if (caracter<32 || caracter>127) {
                               ^

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>